feat: Add Bid And Bidding Alert to service

+++ core-server/src/main/java/com/carrot/core/bid/application/dto/BidRequest.java
@@ -5,6 +5,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @BidRequestCheck
@@ -15,7 +16,7 @@ public record BidRequest(
         @NotNull Long auctionId,
         @Min(value = 1_000)
         @Schema(description = "입찰 가격", example = "50000")
-        int biddingPrice,
+        BigDecimal biddingPrice,
         @Schema(description = "경매 종료 일자", type = "string", example = "2024-03-08T00:00:00+0900")
         LocalDateTime biddingTime) {
 }
